Tried the architecture with one convolution layer
But with the learning rate: 0.001 (Keras default) for RMSprop optimizer and one 2D convolutional layer with 16 neurons as output dimension. There were fluctuations (could see the spikes in the learning curves plot) in both val_accuracy and val_loss compared to training accuracy and training loss. The usual suspects in this situation turns out to be:
1. batch_size
2. learning_rate