Validation loss decreases with training loss
We are in the right direction as validation loss shows improvement. At the same time, validation accuracy is fluctuating. So there is room for improvement. Again batch_size pops up as a solution to look into. Another suggestion is a large network but a small dataset.

What's the possible next step?
We had batch_size set to 128 for training data and 32 for validation data. Maybe try to change them to have smaller batch_size and therefore more number of steps_per_epoch?