Model clearly overfits
When reduced the training batch_size from 128 to 64 with a single layer Conv2D with 64 neurons as output dimension. Although with only 64 images in a batch, the training accuracy and loss curves are way better (accuracy reaches close to 1 and loss decreases with a minimum amount of fluctuations), the accuracy and loss curves of the validation data are worse signaling the wrong direction we are headed.

Possible next step:
Let's add more Conv2D layers and compress the images with the max_pooling layers and try to bring the total trainable parameters down.