Validation accuracy and loss have improved
With 16_32_32_16 architecture we see that the validation accuracy improves and validation loss is lesser than that of the training data. Much improved result so far. Very happy. Not to deny the fact that there are some spikes and the improvement in accuracy and loss was not smooth. At the same time, we need to remember the fact that we had fewer images to begin with.